bug fixing

+++ public/js/views/Product/InvoiceOrder/ProductItems.js
@@ -583,7 +583,7 @@ define([
                     cost = $currentEl.find('[data-name="price"] input').val() || '0';
                     quantity = this.quantityRetriver($currentEl);
                     cost = helpers.spaceReplacer(cost);
-                    totalUntax += parseInt(cost);
+                    totalUntax += parseFloat(cost);
                     date = $currentEl.find('.datepicker').text();
                     dates.push(date);
                 }
